# refactor: simplify silence indicators and improve timing