--- a/client/src/components/LyricsVisualizer/LyricsContent.tsx
+++ b/client/src/components/LyricsVisualizer/LyricsContent.tsx
@@ -120,9 +120,9 @@ const LyricsContent: React.FC = () => {
             data-current={isActive ? "true" : "false"}
             data-line-index={index} // Add data attribute for easier detection
             data-line-text={line.text.substring(0, 20)} // Add data attribute for text matching
-            className={`3xl:text-[clamp(1.5rem,5.5vw,7rem)] 4xl:text-[clamp(1.5rem,6vw,8em)] my-3 transform py-2.5 text-center text-[clamp(1.5rem,4.5vw,3rem)] font-normal opacity-50 transition-all duration-300 lg:text-[clamp(1.5rem,5vw,6rem)] ${
+            className={`3xl:text-[clamp(1.5rem,5.5vw,7rem)] 4xl:text-[clamp(1.5rem,6vw,8em)] my-3 py-2.5 text-center text-[clamp(1.5rem,4.5vw,3rem)] font-normal opacity-50 transition-all duration-300 lg:text-[clamp(1.5rem,5vw,6rem)] ${
               isActive
-                ? "font-black opacity-100 [letter-spacing:0.02em] [text-shadow:0_0_15px_#fff,0_0_30px_#fff,2px_2px_4px_rgba(0,0,0,0.8)]"
+                ? "scale-105 font-black opacity-100 [letter-spacing:0.02em] [text-shadow:0_0_15px_#fff,0_0_30px_#fff,2px_2px_4px_rgba(0,0,0,0.8)]"
                 : ""
             }`}
           >