… CLAUDE.md
… README.md
… client/src/components/LyricsVisualizer/SilenceIndicator.tsx
… client/src/components/LyricsVisualizer/SilenceLine.tsx
… client/src/components/Player/PlayerControls.tsx
… client/src/constants/timing.ts
… client/src/utils/lyricsNormalizer.ts
… client/src/utils/scrollUtils.ts
… client/tests/e2e/functional/settings.spec.ts
… client/tests/unit/lyricsNormalizer.test.ts
… client/tests/unit/silenceIndicator.test.ts