## Summary

  This PR simplifies the silence indicator implementation and improves timing constants for better UX.

  ## Changes

  ### Refactoring
  - **Simplified silence indicator behavior**: Removed delayed visibility logic - indicators now show immediately when active
  - **Cleaned up unused code**: Removed animation callbacks (onEntranceComplete/onExitComplete) and pendingScrollTargetRef
  - **Streamlined scroll logic**: Only handles regular lyric lines now

  ### Timing Improvements
  - **Adjusted DETECTION_THRESHOLD**: 20s → 15s (earlier silence detection)
  - **Adjusted INDICATOR_DELAY**: 1s → 5s (better UX timing)
  - **Removed INDICATOR_END_DELAY**: No longer needed with simplified approach
  - **Single silence marker**: Changed from ♪/♪♪ pair to single ♪ per gap

  ### Testing
  - Updated unit tests to match new timing constants (15s detection, 5s delay)
  - Fixed E2E test race condition in viewport iteration test
  - All tests passing ✅

  ## Impact
  - Better silence detection timing (not too aggressive, not too late)
  - Cleaner, more maintainable codebase
  - More robust E2E tests

  🤖 Generated with [Claude Code](https://claude.com/claude-code)